% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMRCA.R
\name{dMRCA}
\alias{dMRCA}
\title{Density function of T_MRCA.}
\usage{
dMRCA(n, u)
}
\arguments{
\item{n}{How many individuals}

\item{u}{The quantile.}
}
\value{
The density in u
}
\description{
Density function of T_MRCA.
}
\examples{
n = c(2,5,10,100)
col = rainbow(length(n))
for(i in 1:length(n)){
  curve(dMRCA(n = n[i], x), 0, 4, col = col[i], add = i!=1, ylim= c(0,0.9))
}
legend(2.5,1,paste("n = ",n, sep = ""), col=col, lty=1)
}
