% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dptd.R
\name{dptd}
\alias{dptd}
\title{Density function for Phase-Type distribution}
\usage{
dptd(u, pi, S)
}
\arguments{
\item{u}{Vector of values}

\item{pi}{Initial distribution}

\item{S}{Subintensitymatrix}
}
\value{
Returns the density.
}
\description{
Density function for Phase-Type distribution
}
\examples{
curve(dptd(x, 1, -2))
}
