% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dTotal.R
\name{dTotal}
\alias{dTotal}
\title{Density function of T_Total.}
\usage{
dTotal(n, u)
}
\arguments{
\item{n}{number of individuals}

\item{u}{the quantile}
}
\value{
the density in u
}
\description{
Density function of T_Total.
}
\examples{
n = c(2, 5, 10, 20, 50, 100)
col = rainbow(length(n))
for(i in 1:length(n)){
  curve(sapply(x, function(y){
    dTotal(n = n[i], y)}), 0, 14, col = col[i], add = i!=1)
}
}
