% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ired.R
\name{ired}
\alias{ired}
\title{Check if phase-type is irreducible}
\usage{
ired(pi, S)
}
\arguments{
\item{pi}{Innitial distribution}

\item{S}{Subintensity matrix}
}
\value{
TRUE or FALSE
}
\description{
Check if phase-type is irreducible
}
