% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanMRCA.R
\name{MeanMRCA}
\alias{MeanMRCA}
\title{Mean of time from most recent common ancestor of n individuals}
\usage{
MeanMRCA(n)
}
\arguments{
\item{n}{Amount of individuals}
}
\value{
the mean value of the phase-type.
}
\description{
Mean of time from most recent common ancestor of n individuals
}
\examples{
plot(2:20, sapply(2:20, MeanMRCA), type = "l", ylim = c(0,7))
}
