% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanTotal.R
\name{MeanTotal}
\alias{MeanTotal}
\title{Mean of T_total of n individuals}
\usage{
MeanTotal(n)
}
\arguments{
\item{n}{amount of individuals}
}
\value{
the mean
}
\description{
Mean of T_total of n individuals
}
\examples{
plot(2:20, sapply(2:20, MeanTotal), type = "l", ylim = c(0,7))
}
