% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mptd.R
\name{mptd}
\alias{mptd}
\title{Moment of phase-type distribution}
\usage{
mptd(pi, S, order)
}
\arguments{
\item{pi}{Initial distribution}

\item{S}{Subintensity matrix}

\item{order}{Order of moment wanted (1 for mean)}
}
\value{
The calculated moment
}
\description{
Moment of phase-type distribution
}
