% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reward.R
\name{reward}
\alias{reward}
\title{Transformation using rewards}
\usage{
reward(r, pi, S)
}
\arguments{
\item{r}{vector of rewards}

\item{pi}{Initial distribution}

\item{S}{Subintensity matrix}
}
\value{
A list containing the following components:
\itemize{
  \item \code{pi} - New Initial distribution
  \item \code{S} - New subintensity matrix
  \item \code{prob} - Probability of entering states with positive reward before absorption.
}
}
\description{
Transformation using rewards
}
\examples{
S = matrix(c(-6, 6, 0, 0,
              0, -3, 1, 2,
              0, 0, -1, 0,
              0, 0, 0, -1), ncol = 4, byrow = TRUE)
p = c(1, 0, 0, 0)
reward(c(4, 2, 0, 1), p, S) #Counting singles
reward(c(0, 1, 2, 0) , p, S) #Counting doubles
reward(c(0, 0, 0, 1) , p, S) #Counting triples

temp = reward(c(0,1,2,0) , p, S)
hist(rptd(1000, temp$pi, temp$S), breaks = 75, prob = TRUE)
curve(dptd(x,temp$pi, temp$S), add = TRUE)
}
