% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptd.R
\name{rptd}
\alias{rptd}
\title{Simulation from phase-type distribution}
\usage{
rptd(n, pi, S)
}
\arguments{
\item{n}{number of observations}

\item{pi}{Initial distribution}

\item{S}{Subintensity matrix}
}
\value{
vector of absorption times
}
\description{
Simulation from phase-type distribution
}
\examples{
m = 5
pi = c(1, rep(0, m-2))
lambdas = sapply(2:m, function(x){choose(x,2)/x})
S = matrix(append(sapply(lambdas[-(m-1)], function(x){
    c(-x, x, rep(0, m-2))}), -lambdas[m-1]),
   nrow = m-1, byrow = TRUE)
hist(rptd(10^3, pi, S), breaks = 20, freq = FALSE)
curve(dptd(x, pi, S), add = TRUE)
}
