% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarMRCA.R
\name{VarMRCA}
\alias{VarMRCA}
\title{Variance of time from most recent common ancestor of n individuals}
\usage{
VarMRCA(n)
}
\arguments{
\item{n}{Amount of individuals}
}
\value{
the variance value of the phase-type.
}
\description{
Variance of time from most recent common ancestor of n individuals
}
\examples{
VarMRCA(4:5)
}
