% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarTotal.R
\name{VarTotal}
\alias{VarTotal}
\title{Variance of total brachlength from most recent common ancestor of n individuals}
\usage{
VarTotal(n)
}
\arguments{
\item{n}{Amount of individuals}
}
\value{
the variance value of the phase-type.
}
\description{
Variance of total brachlength from most recent common ancestor of n individuals
}
\examples{
VarTotal(4:5)
}
